requirements.txt
streamlit
pandas
matplotlib
wordcloud
textblob
streamlit_lottie
requests
pillow
plotly
streamlit_extras